% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omiqGTR.R
\name{gate}
\alias{gate}
\title{Apply OMIQ gates to cytometry data}
\usage{
gate(gt, data, omiq_id = NULL, sep = "/", ...)
}
\arguments{
\item{gt}{A \code{GatingTree} object created by \code{parse_omiqgt()} \strong{OR} path to an
\code{.omiqgt} file.}

\item{data}{Numeric matrix with named columns (expression matrix, columns
named after fluorophores/channels/markers) \strong{OR} a
\code{flowCore::flowFrame} object \strong{OR} path to an FCS file.}

\item{omiq_id}{Optional sample \emph{OmiqID} if there are any per-file gate
adjustments. Defaults to \code{NULL} (no adjustents).}

\item{sep}{Separator for the path column names. Defaults to "/".}

\item{...}{Additional arguments to \code{flowCore::read.FCS()} if a path to FCS
file was given.}
}
\value{
A logical matrix: one row per cell, one column per path through the
tree (root -> node), TRUE where the cell falls in that population. This is
analogous to \code{flowWorkspace::gh_pop_get_indices()}.
}
\description{
Uses a gating tree exported from OMIQ (as an \code{.omiqgt} file) to gate an FCS
file. If there are any adjustments to gates per individual file, the file's
\emph{OmiqID} must be specified. Otherwise, any potential adjustments will be
ignored, and the default gate definitions will always be applied.
}
