% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omiqGTR.R
\name{plot.GatingTree}
\alias{plot.GatingTree}
\title{Visualize a parsed OMIQ gating tree object}
\usage{
\method{plot}{GatingTree}(
  x,
  horizontal = TRUE,
  cex = 0.62,
  node.cex = 1,
  align.leaves = TRUE,
  edge.col = "grey65",
  edge.lwd = 1,
  leaf.col = "black",
  node.col = "#1f6feb",
  point.pch = 21,
  point.cex = 0.7,
  point.bg = "white",
  main = "OMIQ gating hierarchy",
  mar = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{GatingTree} object created by \code{parse_omiqgt()}.}

\item{horizontal}{TRUE (default) = root at left, populations read top-down.
FALSE = root at top, leaves along the bottom.}

\item{cex}{Base label size (leaves). Internal labels use cex*node.cex.
Defaults to 0.62.}

\item{node.cex}{Relative size of internal-node labels. Defaults to 1.}

\item{align.leaves}{Align leaf labels in a column with dotted leader lines?
Defaults to TRUE.}

\item{edge.col, edge.lwd}{Branch color / width. Default to `"grey65"`` and 1.}

\item{leaf.col, node.col}{Label colours for leaves / internal nodes. Default to
\code{"black"} and \code{"#1f6feb"} (blue).}

\item{point.pch, point.cex, point.bg}{Node markers (set \code{point.pch = NA} to
hide). Default to 21, 0.7, and \code{"white"}.}

\item{main, mar}{Plot title and margins. Default to \emph{"OMIQ gating hierarchy"}
and \code{NULL} (default margins).}

\item{...}{Extra arguments passed to \code{graphics::par()}.}
}
\description{
Draws a dendrogram of gates as defined by a \code{GatingTree} object generated
using the \code{omiqGTR::parse_omiqgt()} function.
}
\details{
\code{GatingTree} object
generated using the \code{omiqGTR::parse_omiqgt()} function.
}
