% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omiqGTR.R
\name{tree_paths}
\alias{tree_paths}
\title{Extract all paths through a parsed OMIQ gating tree object}
\usage{
tree_paths(gt, sep = "|")
}
\arguments{
\item{gt}{A \code{GatingTree} object created by \code{parse_omiqgt()}.}

\item{sep}{Separator for the paths. Defaults to "|".}
}
\value{
A character vector of full node paths.
}
\description{
Returns every full, unique gate path in a \code{GatingTree}. These are identical
to (and in the same order as) the column names of the matrix returned by
\code{omiqGTR::gate()}, provided that the same separator is used.
}
